:recycle: moved GridSchema -> `editgridschema.py` added `coerce_data` fn
